Add ItemGroup for publish profiles and service dependencies
The changes in the `AspNet.Web.Api.csproj` file include the addition of a new `ItemGroup` element. This new `ItemGroup` contains two `Folder` elements that specify the inclusion of the `Properties\PublishProfiles\` and `Properties\ServiceDependencies\` directories. These folders are likely being added to organize publish profiles and service dependencies for the project.